requirements.txt
requests
flask
pillow
brother_ql
qrcode
